+++ HISTORY.md
@@ -2,6 +2,7 @@
 ==================
 
   * Add support ```selection.classed```.  [Bug 57](https://github.com/mhemesath/r2d3/issues/57)
+  * Handle R2D3Elements in insertBefore. [Bug 84](https://github.com/mhemesath/r2d3/issues/84)
 
 0.1.1 / 2013-03-18
 ==================